rabbit: Change start_apps/2 to call application:ensure_all_started/2
.. instead of reinventing about the same logic.

start_apps/2 is not called to start the `rabbit` application anymore (it
is started as a normal Erlang application now). It means it's only used
to start plugins which are also supposed to be regular Erlang
applications.

It run boot steps unconditionally as well (because this function never
starts `rabbit).

One difference with the previous code is that the previous code started
all loaded applications. This seemed unnecessary and even caused issues
with testsuites where some applications were loaded like `rabbitmqctl`
but were not meant to be started.

[#170870798]